+++ frontend/src/CoreConstants.ts
