+++ frontend/src/components/forms/TextInputComponent.vue
@@ -23,7 +23,7 @@
 
 <script setup lang="ts">
 import { ref, watch } from 'vue'
-import { isEmpty } from '@/core/CommonTypesDto'
+import { isEmpty } from '@/dto/CommonTypesDto'
 
 //Define the input properties for this component
 const props = defineProps<{
