+++ frontend/src/composables/useFetch.ts
@@ -1,6 +1,6 @@
 import { ref, type Ref } from 'vue'
 import axios, { type AxiosError } from 'axios'
-import { backendUrl, frontendUrl } from '@/core/CoreConstants'
+import { backendUrl, frontendUrl } from '@/CoreConstants'
 
 axios.defaults.headers.common['Access-Control-Allow-Origin'] = frontendUrl
