+++ frontend/src/dto/CommonTypesDto.ts
