+++ frontend/src/pages/applyform/BusinessInformationWizardStep.vue
@@ -2,10 +2,7 @@
 import { watch, computed, ref, reactive } from 'vue'
 import { useEventBus } from '@vueuse/core'
 import { useFetchTo } from '@/composables/useFetch'
-import {
-  type BusinessSearchResult,
-  ClientTypeEnum
-} from '@/core/CommonTypesDto'
+import { type BusinessSearchResult, ClientTypeEnum } from '@/dto/CommonTypesDto'
 import type {
   FormDataDto,
   ForestClientDetailsDto
