+++ frontend/src/pages/applyform/ContactWizardStep.vue
@@ -56,7 +56,7 @@ import {
 } from '@/dto/ApplyClientNumberDto'
 import { useFetchTo } from '@/composables/useFetch'
 
-import type { CodeNameType, ModalNotification } from '@/core/CommonTypesDto'
+import type { CodeNameType, ModalNotification } from '@/dto/CommonTypesDto'
 
 import Note from '@/common/NoteComponent.vue'
