+++ frontend/src/services/ForestClientService.ts
@@ -1,5 +1,5 @@
 import type { Address, Contact } from '../dto/ApplyClientNumberDto'
-import type { CodeDescrType } from '@/core/CommonTypesDto'
+import type { CodeDescrType } from '@/dto/CommonTypesDto'
 
 export const addNewAddress = (addresses: Address[]): number => {
   const blankAddress: Address = {
